requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
discord.py>=2.3.2
python-dotenv>=1.0.0 